+++ src/main/scala/trw/dbsubsetter/datacopy/DataCopierFactoryImpl.scala
@@ -1,14 +1,14 @@
 package trw.dbsubsetter.datacopy
 
-import trw.dbsubsetter.datacopy.impl.{GenericDataCopierImpl, PostgresDataCopierImpl}
 import trw.dbsubsetter.db.{DbAccessFactory, DbVendor, SchemaInfo}
 
-final class DataCopierFactoryImpl(dbAccessFactory: DbAccessFactory, schemaInfo: SchemaInfo) extends DataCopierFactory {
+private[datacopy] final class DataCopierFactoryImpl(dbAccessFactory: DbAccessFactory, schemaInfo: SchemaInfo)
+    extends DataCopierFactory {
 
   def build(): DataCopier = {
     dbAccessFactory.getDbVendor() match {
-      case DbVendor.PostgreSQL => new PostgresDataCopierImpl(dbAccessFactory, schemaInfo)
-      case _                   => new GenericDataCopierImpl(dbAccessFactory)
+      case DbVendor.PostgreSQL => new DataCopierPostgresImpl(dbAccessFactory, schemaInfo)
+      case _                   => new DataCopierGenericImpl(dbAccessFactory)
     }
   }
 }
