+++ src/main/scala/trw/dbsubsetter/datacopy/DataCopyQueueImpl.scala
@@ -1,8 +1,6 @@
-package trw.dbsubsetter.datacopyqueue.impl
+package trw.dbsubsetter.datacopy
 
 import trw.dbsubsetter.config.Config
-import trw.dbsubsetter.datacopy.DataCopyTask
-import trw.dbsubsetter.datacopyqueue.DataCopyQueue
 import trw.dbsubsetter.db.ColumnTypes.ColumnType
 import trw.dbsubsetter.db.{Constants, Keys, PrimaryKeyValue, SchemaInfo, Table}
 import trw.dbsubsetter.keyextraction.KeyExtractionUtil
@@ -13,7 +11,7 @@ import scala.collection.mutable
 /**
   * WARNING: this class is not threadsafe
   */
-private[datacopyqueue] final class DataCopyQueueImpl(config: Config, schemaInfo: SchemaInfo) extends DataCopyQueue {
+private[datacopy] final class DataCopyQueueImpl(config: Config, schemaInfo: SchemaInfo) extends DataCopyQueue {
 
   private[this] val tablesToQueuedValueCounts: mutable.Map[Table, Long] = {
     val elems: Seq[(Table, Long)] =
