+++ src/main/scala/trw/dbsubsetter/datacopy/DataCopyQueueInstrumented.scala
@@ -1,11 +1,9 @@
-package trw.dbsubsetter.datacopyqueue.impl
+package trw.dbsubsetter.datacopy
 
-import trw.dbsubsetter.datacopy.DataCopyTask
-import trw.dbsubsetter.datacopyqueue.DataCopyQueue
 import trw.dbsubsetter.metrics.Metrics
 import trw.dbsubsetter.pkstore.PksAdded
 
-private[datacopyqueue] final class DataCopyQueueInstrumented(delegatee: DataCopyQueue) extends DataCopyQueue {
+private[datacopy] final class DataCopyQueueInstrumented(delegatee: DataCopyQueue) extends DataCopyQueue {
 
   private[this] val pendingTaskCount = Metrics.PendingDataCopyRows
