+++ src/main/scala/trw/dbsubsetter/fkcalc/FkExtractor.scala
@@ -1,0 +1,20 @@
+package trw.dbsubsetter.fkcalc
+
+import trw.dbsubsetter.db.{ForeignKey, ForeignKeyValue, Keys, SchemaInfo}
+
+private[fkcalc] object FkExtractor {
+  def fkExtractionFunctions(schemaInfo: SchemaInfo): Map[(ForeignKey, Boolean), Keys => ForeignKeyValue] = {
+    schemaInfo.fksOrdered.flatMap { foreignKey =>
+      val parentExtractionFunction: Keys => ForeignKeyValue =
+        keys => keys.getValue(foreignKey, confusingTechDebt = false)
+
+      val childExtractionFunction: Keys => ForeignKeyValue =
+        keys => keys.getValue(foreignKey, confusingTechDebt = true)
+
+      Seq(
+        (foreignKey, false) -> parentExtractionFunction,
+        (foreignKey, true) -> childExtractionFunction
+      )
+    }.toMap
+  }
+}
