+++ src/main/scala/trw/dbsubsetter/fktaskqueue/ForeignKeyTaskQueueFactory.scala
@@ -2,7 +2,6 @@ package trw.dbsubsetter.fktaskqueue
 
 import trw.dbsubsetter.config.Config
 import trw.dbsubsetter.db.SchemaInfo
-import trw.dbsubsetter.fktaskqueue.impl.{ForeignKeyTaskChronicleQueue, ForeignKeyTaskQueueInstrumented}
 
 object ForeignKeyTaskQueueFactory {
   def build(config: Config, schemaInfo: SchemaInfo): ForeignKeyTaskQueue = {
