+++ src/main/scala/trw/dbsubsetter/pkstore/PrimaryKeyStoreFactory.scala
@@ -1,7 +1,6 @@
 package trw.dbsubsetter.pkstore
 
 import trw.dbsubsetter.db.SchemaInfo
-import trw.dbsubsetter.pkstore.impl.{InMemoryPrimaryKeyStore, InstrumentedPrimaryKeyStore}
 
 object PrimaryKeyStoreFactory {
 
@@ -9,7 +8,7 @@ object PrimaryKeyStoreFactory {
    * Only call this once per subset (it needs to be a singleton per subsetting run because it holds state)
    */
   def buildPrimaryKeyStore(schemaInfo: SchemaInfo): PrimaryKeyStore = {
-    val base: PrimaryKeyStore = new InMemoryPrimaryKeyStore(schemaInfo)
-    new InstrumentedPrimaryKeyStore(base)
+    val base: PrimaryKeyStore = new PrimaryKeyStoreInMemoryImpl(schemaInfo)
+    new PrimaryKeyStoreInstrumentedImpl(base)
   }
 }
