+++ src/main/scala/trw/dbsubsetter/pkstore/PrimaryKeyStoreInMemoryImpl.scala
@@ -1,14 +1,13 @@
-package trw.dbsubsetter.pkstore.impl
+package trw.dbsubsetter.pkstore
 
 import trw.dbsubsetter.db.{PrimaryKeyValue, SchemaInfo, Table}
-import trw.dbsubsetter.pkstore._
 
 import scala.collection.mutable
 
 /**
   * WARNING: this class is not threadsafe
   */
-private[pkstore] final class InMemoryPrimaryKeyStore(schemaInfo: SchemaInfo) extends PrimaryKeyStore {
+private[pkstore] final class PrimaryKeyStoreInMemoryImpl(schemaInfo: SchemaInfo) extends PrimaryKeyStore {
 
   /*
    * If `seenWithChildrenStorage` contains a PK, then both its children AND its parents have been fetched.
@@ -19,13 +18,13 @@ private[pkstore] final class InMemoryPrimaryKeyStore(schemaInfo: SchemaInfo) ext
    * `seenWithChildrenStorage` -- it will never be in both at once.
    */
   private[this] val seenWithoutChildrenStorage: Map[Table, mutable.HashSet[Any]] =
-    InMemoryPrimaryKeyStore.buildStorage(schemaInfo)
+    PrimaryKeyStoreInMemoryImpl.buildStorage(schemaInfo)
 
   private[this] val seenWithChildrenStorage: Map[Table, mutable.HashSet[Any]] =
-    InMemoryPrimaryKeyStore.buildStorage(schemaInfo)
+    PrimaryKeyStoreInMemoryImpl.buildStorage(schemaInfo)
 
   override def markSeen(table: Table, primaryKeyValue: PrimaryKeyValue): WriteOutcome = {
-    val rawValue: Any = InMemoryPrimaryKeyStore.extract(primaryKeyValue)
+    val rawValue: Any = PrimaryKeyStoreInMemoryImpl.extract(primaryKeyValue)
 
     val alreadySeenWithChildren: Boolean =
       seenWithChildrenStorage(table).contains(rawValue)
@@ -44,7 +43,7 @@ private[pkstore] final class InMemoryPrimaryKeyStore(schemaInfo: SchemaInfo) ext
   }
 
   override def markSeenWithChildren(table: Table, primaryKeyValue: PrimaryKeyValue): WriteOutcome = {
-    val rawValue: Any = InMemoryPrimaryKeyStore.extract(primaryKeyValue)
+    val rawValue: Any = PrimaryKeyStoreInMemoryImpl.extract(primaryKeyValue)
 
     val alreadySeenWithChildren: Boolean =
       !seenWithChildrenStorage(table).add(rawValue)
@@ -63,12 +62,12 @@ private[pkstore] final class InMemoryPrimaryKeyStore(schemaInfo: SchemaInfo) ext
   }
 
   override def alreadySeen(table: Table, primaryKeyValue: PrimaryKeyValue): Boolean = {
-    val rawValue: Any = InMemoryPrimaryKeyStore.extract(primaryKeyValue)
+    val rawValue: Any = PrimaryKeyStoreInMemoryImpl.extract(primaryKeyValue)
     seenWithChildrenStorage(table).contains(rawValue) || seenWithoutChildrenStorage(table).contains(rawValue)
   }
 }
 
-private object InMemoryPrimaryKeyStore {
+private object PrimaryKeyStoreInMemoryImpl {
   private def buildStorage(schemaInfo: SchemaInfo): Map[Table, mutable.HashSet[Any]] = {
     val tables: Iterable[Table] = schemaInfo.pksByTable.keys
     tables.map { t => t -> mutable.HashSet.empty[Any] }.toMap
